Move tag interfaces to base
