+++ android/base/src/main/kotlin/net/twisterrob/inventory/android/arch/Mapper.kt
@@ -1,0 +1,5 @@
+package net.twisterrob.inventory.android.arch
+
+fun interface Mapper<Input, Output> {
+	fun map(input: Input): Output
+}
