+++ android/base/src/main/kotlin/net/twisterrob/inventory/android/arch/UiEffect.kt
@@ -1,0 +1,3 @@
+package net.twisterrob.inventory.android.arch
+
+interface UiEffect
