+++ android/base/src/main/kotlin/net/twisterrob/inventory/android/arch/UiEffectHandler.kt
@@ -1,0 +1,5 @@
+package net.twisterrob.inventory.android.arch
+
+fun interface UiEffectHandler<Effect : UiEffect> {
+	suspend fun launch(effect: Effect)
+}
