+++ android/base/src/main/kotlin/net/twisterrob/inventory/android/arch/UiEventHandler.kt
@@ -1,0 +1,5 @@
+package net.twisterrob.inventory.android.arch
+
+fun interface UiEventHandler {
+	fun wireEvents()
+}
