+++ android/base/src/main/kotlin/net/twisterrob/inventory/android/arch/UiStateHandler.kt
@@ -1,0 +1,5 @@
+package net.twisterrob.inventory.android.arch
+
+fun interface UiStateHandler<State : UiState> {
+	fun render(state: State)
+}
