+++ android/base/src/main/kotlin/net/twisterrob/inventory/android/arch/UseCase.kt
@@ -1,0 +1,5 @@
+package net.twisterrob.inventory.android.arch
+
+fun interface UseCase<Input, Output> {
+	suspend fun execute(input: Input): Output
+}
